+++ lib/calendar_assistant/cli_helpers.rb
@@ -1,16 +1,18 @@
 # coding: utf-8
 class CalendarAssistant
   module CLIHelpers
+    class ChronicParseException < CalendarAssistant::BaseException ; end
+
     def self.parse_datespec userspec
       start_userspec, end_userspec = userspec.split(/ ?\.\.\.? ?/)
 
       if end_userspec.nil?
-        time = Chronic.parse(userspec) || raise("could not parse #{userspec}")
+        time = Chronic.parse(userspec) || raise(ChronicParseException, "could not parse '#{userspec}'")
         return time.beginning_of_day..time.end_of_day
       end
 
-      start_time = Chronic.parse(start_userspec) || raise("could not parse #{start_userspec}")
-      end_time = Chronic.parse(end_userspec) || raise("could not parse #{end_userspec}")
+      start_time = Chronic.parse(start_userspec) || raise(ChronicParseException, "could not parse '#{start_userspec}'")
+      end_time = Chronic.parse(end_userspec) || raise(ChronicParseException, "could not parse '#{end_userspec}'")
 
       if start_time.to_date == end_time.to_date
         start_time..end_time